fix: Move methods to a Tenant Repo
Attempt at moving some of the existing methods to a new module focused on Tenants

This module won't use Ecto.Repo since it has the intent to use mostly Postgrex queries directly
