+++ lib/realtime/tenants/tenant_repo.ex
@@ -1,0 +1,154 @@
+defmodule Realtime.Tenants.TenantRepo do
+  require Logger
+
+  import Ecto.Query
+  alias Ecto.Repo.Schema
+  alias Ecto.Adapters.Postgres
+
+  @doc """
+  Lists all records for a given query and converts them into a given struct
+  """
+  @spec all(DBConnection.conn(), Ecto.Queryable.t(), module(), [Postgrex.execute_option()]) ::
+          {:ok, list(struct())} | {:error, any()}
+  def all(conn, query, result_struct, opts \\ []) do
+    conn
+    |> run_all_query(query, opts)
+    |> result_to_structs(result_struct)
+  end
+
+  @doc """
+  Fetches one record for a given query and converts it into a given struct
+  """
+  @spec one(DBConnection.conn(), Ecto.Query.t(), module()) ::
+          {:ok, struct()} | {:ok, nil} | {:error, any()}
+  def one(conn, query, result_struct) do
+    conn
+    |> run_all_query(query)
+    |> result_to_single_struct(result_struct)
+  end
+
+  @doc """
+  Inserts a given changeset into the database and converts the result into a given struct
+  """
+  @spec insert(DBConnection.conn(), Ecto.Changeset.t(), module()) ::
+          {:ok, struct()} | {:error, any()} | Ecto.Changeset.t()
+  def insert(conn, changeset, result_struct) do
+    with {:ok, {query, args}} <- insert_query_from_changeset(changeset) do
+      conn
+      |> run_query_with_trap(query, args)
+      |> result_to_single_struct(result_struct)
+    end
+  end
+
+  @doc """
+  Deletes records for a given query and returns the number of deleted records
+  """
+  @spec del(DBConnection.conn(), Ecto.Queryable.t()) ::
+          {:ok, non_neg_integer()} | {:error, any()}
+  def del(conn, query) do
+    with {:ok, %Postgrex.Result{num_rows: num_rows}} <- run_delete_query(conn, query) do
+      {:ok, num_rows}
+    end
+  end
+
+  @doc """
+  Updates an entry based on the changeset and returns the updated entry
+  """
+  @spec update(DBConnection.conn(), Ecto.Changeset.t(), module()) ::
+          {:ok, struct()} | {:error, any()} | Ecto.Changeset.t()
+  def update(conn, changeset, result_struct, opts \\ []) do
+    with {:ok, {query, args}} <- update_query_from_changeset(changeset) do
+      conn
+      |> run_query_with_trap(query, args, opts)
+      |> result_to_single_struct(result_struct)
+    end
+  end
+
+  defp result_to_single_struct({:error, _} = error, _), do: error
+
+  defp result_to_single_struct({:ok, %Postgrex.Result{rows: []}}, _), do: {:error, :not_found}
+
+  defp result_to_single_struct({:ok, %Postgrex.Result{rows: [row], columns: columns}}, struct) do
+    {:ok, Schema.load(Postgres, struct, Enum.zip(columns, row))}
+  end
+
+  defp result_to_single_struct({:ok, %Postgrex.Result{num_rows: num_rows}}, _) do
+    raise("expected at most one result but got #{num_rows} in result")
+  end
+
+  defp result_to_structs({:error, _} = error, _), do: error
+
+  defp result_to_structs({:ok, %Postgrex.Result{rows: rows, columns: columns}}, struct) do
+    {:ok, Enum.map(rows, &Schema.load(Postgres, struct, Enum.zip(columns, &1)))}
+  end
+
+  defp insert_query_from_changeset(%{valid?: false} = changeset), do: {:error, changeset}
+
+  defp insert_query_from_changeset(changeset) do
+    schema = changeset.data.__struct__
+    source = schema.__schema__(:source)
+    prefix = schema.__schema__(:prefix)
+    acc = %{header: [], rows: []}
+
+    %{header: header, rows: rows} =
+      Enum.reduce(changeset.changes, acc, fn {field, row}, %{header: header, rows: rows} ->
+        %{
+          header: [Atom.to_string(field) | header],
+          rows: [row | rows]
+        }
+      end)
+
+    table = "\"#{prefix}\".\"#{source}\""
+    header = "(#{header |> Enum.map(&"\"#{&1}\"") |> Enum.join(",")})"
+
+    arg_index =
+      rows
+      |> Enum.with_index(1)
+      |> Enum.map(fn {_, index} -> "$#{index}" end)
+      |> Enum.join(",")
+
+    {:ok, {"INSERT INTO #{table} #{header} VALUES (#{arg_index}) RETURNING *", rows}}
+  end
+
+  defp update_query_from_changeset(%{valid?: false} = changeset), do: {:error, changeset}
+
+  defp update_query_from_changeset(changeset) do
+    %Ecto.Changeset{data: %{id: id, __struct__: struct}, changes: changes} = changeset
+    changes = Keyword.new(changes)
+    query = from(c in struct, where: c.id == ^id, select: c, update: [set: ^changes])
+    {:ok, to_sql(:update_all, query)}
+  end
+
+  defp run_all_query(conn, query, opts \\ []) do
+    {query, args} = to_sql(:all, query)
+    run_query_with_trap(conn, query, args, opts)
+  end
+
+  defp run_delete_query(conn, query) do
+    {query, args} = to_sql(:delete_all, query)
+    run_query_with_trap(conn, query, args)
+  end
+
+  defp run_query_with_trap(conn, query, args, opts \\ []) do
+    Postgrex.query(conn, query, args, opts)
+  rescue
+    exception ->
+      Logger.error("Postgrex exception: #{inspect(exception)}")
+      {:error, :postgrex_exception}
+  catch
+    :exit, reason ->
+      Logger.error("Postgrex exit: #{inspect(reason)}")
+
+      {:error, :postgrex_exception}
+  end
+
+  defp to_sql(operation, queryable) do
+    {_, {_, _, {_, prepared}}, dump_params} =
+      queryable
+      |> Ecto.Queryable.to_query()
+      |> Ecto.Query.Planner.ensure_select(operation == :all)
+      |> Ecto.Query.Planner.query(operation, __MODULE__, Postgres, 0)
+
+    {prepared, dump_params}
+  end
+end
